server/requirements.txt
fastapi
uvicorn
python-socketio
motor
pymongo
python-dotenv
dnspython
